vmm: make output for failed vmexits more threadsafe
If you had concurrent threads vmexit and the 2LS was unable to handle
the exits, you'd get an unintelligible mess.

This commit removes some unnecessary info - namely all of the info in
handle_vmexit() - and protects all of the output with a single lock.
It's not super elegant, and other printing can interfere with the
output, but it helps when there is a coordinated failure across numerous
threads.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>