9ns: make the diagnostic output of bind saner
The biggest issue was the X -> Y output; it was the opposite of what you
see from a ls -l from a symlink.  The arrow had meant "put X onto Y",
but in symlink terms, that is "X points to Y" which is the opposite.

Plus, you could do something like bind -ac and only the -a was noticed.
Fun.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>