9ns: move the symlink check after the mount check
Although you can't bind a symlink (so that it is the domount side of a
mount point), the code is a bit cleaner if we do the symlink check after
the 'mount/nomount' spot.

Also, if we ever build in support for binding symlinks (src_path / WHAT)
onto the namespace, then we'll already have support for that on the
intermediate path names in a walk.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>