Remove extraneous sysfd2path()
We already have sys_fd2path(), and no one was using sysfd2path().  The
latter, which we are removing, is the Plan 9 implementation, I think.

Arguably, the one we are removing is a little faster, since it doesn't
invoke the snprintf() mechanisms and is a straight memcpy, but it also
doesn't handle the "no string" case yet.  If we ever want to make
sys_fd2path() faster, we certainly can.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>