9ns: don't pass user pointers for 'spec'
Spec is often passed directly to error(), and thus snprintf.  Right now,
we don't actually pass a user pointer spec.  This comment tracks the one
location we could in the future.

Note that when you bind '#dev.spec', that string is copied in to the
kernel and checked.  That's in namec().

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>