mm: remove the path arguments to foc_abs_path()
I think the VFS version of foc_abs_path() needed the buffer for the
storage of the absolute path, since the VFS didn't maintain that
internally.  9ns maintains that pointer, so we can just return it.

The lifetime of that chan's name should be the lifetime of the chan
itself.  Although I can imagine cases where that isn't true, e.g. if you
walk a chan to another chan, hopefully those chans are externalized to
the rest of the system, i.e. only used in chan.c during walk().  Keep in
mind walks from O_PATH chans.

If that turns out to not be the case, then we can copy it out, though
even then we'd need to understand when the chan's name is changing,
since there's no reason it can't be changed during foc_abs_path().

Similarly, I'd like to know if we ever have a chan exposed to the rest
of the system that has no name.  The answer to this is probably related
to whether or not the chan's name can change.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>