9ns: change parsecmd()'s size arg's type to size_t
This is a leftover from Inferno/Plan 9, where ints are scattered about
where we should use a size_t.  Longs are a little better (64 bit vs 32),
but are still signed.

This commit changes parsecmd() and all of its callers.  Beyond the
ether->ctl and whatnot, there probably are other users who still convert
length to an int, which can cause problems.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>