qio: implement pullupblock() for block extra data
mnt tickled this when mounting ufs served from a VM.  We had a
large-enough pullupblock that reached into the ebd.  This required
calling pullupblock on something that wasn't just network data.  #mnt
also did a few things like read just the 9p headers, then read the rest,
which may have caused the issue.  Either that, or it was just a larger
packet.

Anyway, the time has come to fix the block extra data once and for all.
While we're here, I changed it to take a size_t as well.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>