arena: catch bad spans containing '0'
Currently arena's can't hand out the value '0'.  That's actually baked
in to the vmem interface.  NULL is returned on error, so you can't
differentiate between it and 0.  I'm OK with that for now.

The asserts will catch people who try to add a span starting at 0, which
is close enough to catch this problem.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>