vmm: refactor userspace's emsr_fakewrite()
The old fakewrite would attempt to do a readmsr, which will fail.
However, we only need to do a readmsr if we read before writing.  Most
uses of this will do a write first.

Note that rdmsr and wrmsr will fail from userspace.  The MSR emulation
code is mostly just an unused copy of the kernel's, but it's useful to
prototype changes in userspace without requiring a kernel reboot.  i.e.
this is for debugging.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>