pci: rename device->linux_dev
We have dev_id, which is Linux's device.  We had device, which was
Linux's dev.  We also have 'dev', which is encoded in Linux's devfn.
This was a mess, especially since when we spatched dev to device, that
collided with other uses of device (meaning dev_id) in Linux.

Anyway, we aren't even using linux_dev nee device, it was just a source
of confusion for me.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>