alarm: do not wait for unset when resetting an alarm
reset_alarm_* are racy.  Concurrent resetters can both unset, then both
set, and set_alarm can't be called on an alarm that is already set.

Right now, the only thing going for reset_alarm is that it makes sure
the alarm has stopped before we set it again.  That seems to prevent the
race between reset_alarm for the handler and an external function, but
does nothing for the race between two random functions.

But actually, the handlers can't call reset_alarm (prior to this
commit); unset will block on the handler finishing, which is a deadlock.
(cv_wait->can_block will catch this).

Basically, handlers needed to use set_alarm to rearm, since they know
they aren't on the tchain anymore.  Hokey.  set_alarm works outside a
handler too (and not the old __set_alarm vs set_alarm nonsense), but
that's only for when you know a handler isn't armed.

A lot of code wants to use reset_alarm, which is similar to Linux's
mod_timer.  It's better to have all of those functions callable in any
context - particularly the most useful one.

This commit doesn't fix the race between unset and set yet.  Instead,
has us unset without waiting for the handler to finish before setting.
This makes it usable in alarm handlers.  You may then worry - can't this
alarm get rearmed and run on another tchain (cpu) concurrently, and then
we're racing on alarm state?  No - if the alarm was currently running,
it was presumably on the same tchain that we're putting it on -
otherwise the caller was bugged.  If it's on the same tchain, then we're
OK, since we know a given tchain processes alarms in order.

A note on naming issues: nosync vs sync, and what's the default can get
sorted out in a future commit.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>