spatch: add a few features to the Linux coccis
Notably, the lock transform was mismatched.  We'd set all locks to be
irqsave (by the init function), but the _bh locks were not irqsave.

The issue here is that _bh locks *don't* need to be irqsave, but we
don't know when we look at spinlock_init where the lock is grabbed.  At
least not easily.  So since we say all locks are irqsave, we need to be
consistent.

Another option would be to keep them as spin_lock_bh and #define it, so
we keep the info that it is a bh lock.  Or if we really care about a
slightly lower lock in a Linux driver, we just look at the changelog.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>