arena: allow a nocross xalloc() with a source arena
When we need to import segments from a parent arena to satisfy an
xalloc, we need to make sure that a successful import actually satisfies
the xalloc.

This doesn't seem easily solvable for min/maxaddr, short of pushing the
xalloc all the way down the chain or having some other inter-arena
interface.  But it should be solvable for nocross, by getting a large
enough allocation.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>