arena: do not round-up when picking xalloc lists
The list we start at is an optimization.  We could easily start at the
first list.  We skip to the first list that *can* satisfy us.
Previously, due to align and phase acrobatics, we could jump to the next
list beyond the one that could satisfy us, which could lead to an "OOM"
when we actually had the resource.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>