arena: allow "self-sourced" arenas
These arenas generate their own resources dynamically and arbitrarily.
They do not source from another arena, such as base, but they also are
not given their segments in advance with arena_add().  Instead, they
create resources when asked.

It is similar to calling arena_add() in response to a failed allocation,
except the allocation never fails.

The arena alloc function pointer lets you do whatever you want.  The
reason for self-sourcing instead of pulling from a dummy arena is that
the self-source provides a pointer to the arena in the allocation
function.  With that pointer, you now have your original arena, which
could be embedded in another structure.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>