pci: add support for assigning and resetting devices
This support is rather limited.  The intent is to allow devices to be
assigned to a process, to initialize the driver, use it as a process,
and then tear down when the process exits.

Ultimately, the driver's lifecycle is a subset of the time it is
assigned to a process.  Drivers that can support repeatedly
initializing and resetting their device publish this when they tell PCI
about their init/reset ops.

By having the driver only exist when the process exists, we can create
an arena for DMA allocs from the user's address space that will exist
for the life of the driver.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>